+++ packages/yourloops/app/main-lobby.tsx
@@ -27,8 +27,9 @@
 
 import React from "react";
 import { Redirect, Route, Switch, useLocation } from "react-router-dom";
+import { useAuth0 } from "@auth0/auth0-react";
 
-import { ThemeProvider, Theme, makeStyles, useTheme } from "@material-ui/core/styles";
+import { makeStyles, Theme, ThemeProvider, useTheme } from "@material-ui/core/styles";
 import CssBaseline from "@material-ui/core/CssBaseline";
 
 import { SessionTimeout, useAuth } from "../lib/auth";
@@ -38,13 +39,12 @@ import Footer from "../components/footer/footer";
 import PatientConsentPage from "../pages/patient/patient-consent";
 import { ConsentPage, LoginPage } from "../pages/login";
 import { SignUpPage } from "../pages/signup";
-import { ConfirmPasswordResetPage, RequestPasswordResetPage } from "../pages/password-reset";
 import { MainLayout } from "../pages/main-layout";
 import InvalidRoute from "../components/invalid-route";
 
 const RENEW_CONSENT_PATH = "/renew-consent";
 const NEW_CONSENT_PATH = "/new-consent";
-const PUBLIC_ROUTES = ["/login", "/signup", "/request-password-reset", "/confirm-password-reset"];
+const PUBLIC_ROUTES = ["/login", "/signup"];
 const EXTERNAL_THEME_ROUTES = [NEW_CONSENT_PATH, RENEW_CONSENT_PATH, ...PUBLIC_ROUTES];
 
 interface StyleProps {
@@ -71,8 +71,8 @@ const routeStyle = makeStyles<Theme, StyleProps>(() => {
 });
 
 export function MainLobby(): JSX.Element {
-
-  const { isLoggedIn, user, isAuthInProgress, isAuthHookInitialized } = useAuth();
+  const { isLoading, isAuthenticated } = useAuth0();
+  const { user } = useAuth();
   const location = useLocation();
   const currentRoute = location.pathname;
   const isCurrentRoutePublic = PUBLIC_ROUTES.includes(currentRoute);
@@ -85,13 +85,13 @@ export function MainLobby(): JSX.Element {
   const renewConsentPath = currentRoute === RENEW_CONSENT_PATH || currentRoute === NEW_CONSENT_PATH;
   let redirectTo = null;
 
-  if (!isCurrentRoutePublic && (isAuthInProgress || !isAuthHookInitialized)) {
+  if (!isCurrentRoutePublic && isLoading) {
     return <React.Fragment />;
   }
 
-  if (isCurrentRoutePublic && isLoggedIn) {
+  if (isCurrentRoutePublic && isAuthenticated) {
     redirectTo = "/";
-  } else if (!isCurrentRoutePublic && !isLoggedIn) {
+  } else if (!isAuthenticated && !isCurrentRoutePublic) {
     redirectTo = "/login";
   } else if (!renewConsentPath && user && user.isUserPatient() && user.shouldAcceptConsent()) {
     redirectTo = "/new-consent";
@@ -102,26 +102,25 @@ export function MainLobby(): JSX.Element {
   return (
     <React.Fragment>
       {redirectTo ? <Redirect to={redirectTo} /> :
-        <ThemeProvider theme={theme}>
-          <SessionTimeout />
-          <CssBaseline />
-          <SnackbarContextProvider context={DefaultSnackbarContext}>
-            <div className={style}>
-              <Switch>
-                <Route exact path="/login" component={LoginPage} />
-                <Route exact path="/signup" component={SignUpPage} />
-                <Route exact path="/request-password-reset" component={RequestPasswordResetPage} />
-                <Route exact path="/confirm-password-reset" component={ConfirmPasswordResetPage} />
-                <Route exact path="/renew-consent" component={ConsentPage} />
-                <Route exact path="/new-consent" component={PatientConsentPage} />
-                <Route exact path="/not-found" component={InvalidRoute} />
-                <Route component={MainLayout} />
-              </Switch>
-            </div>
-          </SnackbarContextProvider>
-          <Footer />
-        </ThemeProvider>
-      }
+        (!isLoading &&
+          <ThemeProvider theme={theme}>
+            <SessionTimeout />
+            <CssBaseline />
+            <SnackbarContextProvider context={DefaultSnackbarContext}>
+              <div className={style}>
+                <Switch>
+                  <Route exact path="/login" component={LoginPage} />
+                  <Route exact path="/signup" component={SignUpPage} />
+                  <Route exact path="/renew-consent" component={ConsentPage} />
+                  <Route exact path="/new-consent" component={PatientConsentPage} />
+                  <Route exact path="/not-found" component={InvalidRoute} />
+                  <Route component={MainLayout} />
+                </Switch>
+              </div>
+            </SnackbarContextProvider>
+            <Footer />
+          </ThemeProvider>
+        )}
     </React.Fragment>
   );
 }